Introduce Custom watcher support
